requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.6.15
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
ecdsa==0.19.1
fastapi==0.115.13
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
mypy==1.16.1
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.1
pytest==8.4.1
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
watchfiles==1.1.0
websockets==15.0.1
